backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
aiosqlite==0.19.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
PyNaCl==1.5.0
cryptography==46.0.3
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
slowapi==0.1.9
